+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-commons/pom.xml
@@ -13,6 +13,18 @@
 
     <artifactId>spring-cloud-alibaba-commons</artifactId>
     <name>Spring Cloud Alibaba Commons</name>
+    <dependencies>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-web</artifactId>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <optional>true</optional>
+        </dependency>
+    </dependencies>
 
 
 </project>
