+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-governance/governance-core/spring-cloud-starter-alibaba-governance-authentication/auth-core/pom.xml
@@ -34,7 +34,11 @@
 			<groupId>com.alibaba.cloud</groupId>
 			<artifactId>auth-data</artifactId>
 			<version>2.2.9-SNAPSHOT</version>
-			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-autoconfigure</artifactId>
+			<optional>true</optional>
 		</dependency>
 	</dependencies>
