+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-governance/governance-core/spring-cloud-starter-alibaba-governance-authentication/auth-data/src/main/java/com/alibaba/cloud/governance/auth/util/JwtUtil.java
@@ -21,9 +21,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import com.alibaba.cloud.commons.lang.StringUtils;
+import com.alibaba.cloud.commons.pair.Pair;
 import com.alibaba.cloud.governance.auth.rule.JwtRule;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.tuple.Pair;
 import org.jose4j.jwk.JsonWebKeySet;
 import org.jose4j.jws.JsonWebSignature;
 import org.jose4j.jwt.JwtClaims;
