+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-governance/governance-core/spring-cloud-starter-alibaba-governance-labelrouting/label-routing-data/src/main/java/com/alibaba/cloud/router/data/controlplane/ControlPlaneConnection.java
@@ -21,7 +21,6 @@ import java.util.List;
 import com.alibaba.cloud.router.data.crd.UntiedRouteDataStructure;
 import com.alibaba.cloud.router.data.repository.RouteDataRepository;
 
-import org.springframework.beans.factory.annotation.Autowired;
 
 /**
  * @author HH
