+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-governance/resource-transform/istio-resource-transform/src/main/java/com/alibaba/cloud/governance/istio/protocol/impl/RdsProtocol.java
@@ -21,6 +21,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.alibaba.cloud.commons.governance.matcher.MatcherType;
+import com.alibaba.cloud.commons.governance.matcher.StringMatcher;
 import com.alibaba.cloud.router.data.controlplane.ControlPlaneConnection;
 import com.alibaba.cloud.router.data.crd.LabelRouteData;
 import com.alibaba.cloud.router.data.crd.MatchService;
@@ -28,8 +30,7 @@ import com.alibaba.cloud.router.data.crd.UntiedRouteDataStructure;
 import com.alibaba.cloud.router.data.crd.rule.HeaderRule;
 import com.alibaba.cloud.router.data.crd.rule.RouteRule;
 import com.alibaba.cloud.router.data.crd.rule.UrlRule;
-import com.alibaba.cloud.governance.common.matcher.MatcherType;
-import com.alibaba.cloud.governance.common.matcher.StringMatcher;
+
 import com.alibaba.cloud.governance.istio.XdsChannel;
 import com.alibaba.cloud.governance.istio.XdsScheduledThreadPool;
 import com.alibaba.cloud.governance.istio.protocol.AbstractXdsProtocol;
